Paragraph and headings no longer have children
Instead, they have a RawChars() Text array that stores the raw text found in the block (including newlines). This is to facilitate later inline parsing. It has broken lots of the tests however as the phase 1 AST truth needs updating to reflect the fact that soft and hard breaks are no longer created at this stage of parsing.